+++ src/interface/tools.js
@@ -8,12 +8,28 @@ import {Tool} from "../tools";
  * @param {HTMLCanvasElement} canvasElement
  * @param {Tool} tool
  */
-export function activateTool(canvasElement, tool) {
+export function addTool(canvasElement, tool) {
 
   assert.instanceOf(tool, Tool);
 
   const imageContext = getElementImageContext(canvasElement);
   imageContext.addTool(tool);
+
+  canvasElement.dispatchEvent(new CustomEvent('ostrich.toolAdded', {
+    detail: {
+      tool: tool,
+    },
+  }));
+}
+
+/**
+ * @param {HTMLCanvasElement} canvasElement
+ * @param {string} name
+ */
+export function activateTool(canvasElement, name) {
+
+  const imageContext = getElementImageContext(canvasElement);
+  const tool = imageContext.getTool(name);
 
   tool.activate(canvasElement);
 
@@ -34,8 +50,6 @@ export function deactivateTool(canvasElement, name) {
   const tool = imageContext.getTool(name);
 
   tool.deactivate(canvasElement);
-
-  imageContext.removeTool(name);
 
   canvasElement.dispatchEvent(new CustomEvent('ostrich.toolDeactivated', {
     detail: {